inference/requirements-rag.txt
pypdf==4.3.1
python-docx==1.1.2
